+++ src/test/java/io/vntr/hermes/HermesUserTest.java
@@ -17,7 +17,7 @@ public class HermesUserTest {
 
     @Test
     public void testGetLogicalUser() {
-        double gamma = 1.5;
+        float gamma = 1.5f;
         Map<Integer, Set<Integer>> partitions = new HashMap<Integer, Set<Integer>>();
         partitions.put(1, initSet( 1,  2,  3,  4,  5,  6,  7));
         partitions.put(2, initSet( 8,  9, 10, 11, 12, 13, 14));
