+++ src/test/java/io/vntr/sparmes/SparmesTestUtils.java
@@ -12,7 +12,7 @@ import java.util.Set;
  */
 public class SparmesTestUtils {
 
-    public static SparmesManager initGraph(int minNumReplicas, double gamma, int numPartitions, Map<Integer, Set<Integer>> friendships) {
+    public static SparmesManager initGraph(int minNumReplicas, float gamma, int numPartitions, Map<Integer, Set<Integer>> friendships) {
         Set<Integer> pids = new HashSet<Integer>();
         for(int pid = 0; pid < numPartitions; pid++) {
             pids.add(pid);
@@ -22,7 +22,7 @@ public class SparmesTestUtils {
         return initGraph(minNumReplicas, gamma, partitions, friendships, replicas);
     }
 
-    public static SparmesManager initGraph(int minNumReplicas, double gamma, Map<Integer, Set<Integer>> partitions, Map<Integer, Set<Integer>> friendships, Map<Integer, Set<Integer>> replicaPartitions) {
+    public static SparmesManager initGraph(int minNumReplicas, float gamma, Map<Integer, Set<Integer>> partitions, Map<Integer, Set<Integer>> friendships, Map<Integer, Set<Integer>> replicaPartitions) {
         SparmesManager manager = new SparmesManager(minNumReplicas, gamma);
         for(Integer pid : partitions.keySet()) {
             manager.addPartition(pid);
